// src/data/enums/page-features.ts
export enum PageFeatureType {
  LINK = 'LINK',
  HEADER = 'HEADER',
  FOOTER_ICON = 'FOOTER_ICON',
  IOS_APP = 'IOS_APP',
  HUAWEI_APP = 'HUAWEI_APP',
  GALAXY_APP = 'GALAXY_APP',
  ANDROID_APP = 'ANDROID_APP',
  CONTACT_FORM = 'CONTACT_FORM',
  MAIL_TO = 'MAIL_TO', //need to disscuss
  CALL_LINK = 'CALL_LINK',
  SMS_LINK = 'SMS_LINK',
  SMS_SHORT_CODE = 'SMS_SHORT_CODE',
  PAID_MESSAGE = 'PAID_MESSAGE', //need to disscuss
  EVENT = 'EVENT',
  VIDEO = 'VIDEO',
  HTML = 'HTML',
  AUDIO = 'AUDIO',
  APPLE_MUSIC = 'APPLE_MUSIC',
  SPOTIFY = 'SPOTIFY',
  YOUTUBE = 'YOUTUBE',
  AUDIOMACK = 'AUDIOMACK',
  DEEZER = 'DEEZER',
  PANDORA = 'PANDORA',
  BOOMPLAY = 'BOOMPLAY',
  PLAT_X_PAY = 'PLAT_X_PAY', //need discuss how it works
  PRODUCT = 'PRODUCT',
  WHATSAPP = 'WHATSAPP',
  FACEBOOK = 'FACEBOOK',
  LINKEDIN = 'LINKEDIN',
  X = 'X',
  TIKTOK = 'TIKTOK',
  PINTEREST = 'PINTEREST',
  JOIN_TELEGRAM_CHANNEL = 'JOIN_TELEGRAM_CHANNEL',
  JOIN_TO_TWITCH = 'TWITCH',
  JOIN_TO_X_SPACE = 'JOIN_TO_X_SPACE',
  JOIN_TO_SLACK = 'JOIN_TO_SLACK',
  JOIN_TO_CLUBHOUSE = 'JOIN_TO_CLUBHOUSE',
  SUBSCRIBE_TO_NEWS = 'SUBSCRIBE_TO_NEWS', // what is token
  SLIDER = 'SLIDER',
  SOCIAL_HANDLES = 'SOCIAL_HANDLES',
  INSTAGRAM = 'INSTAGRAM',
}